# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
      "Footprint": {
        "description": "A building outline as a ring of lat/lng vertices (source: OSM).",
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/LatLng"
            },
            "title": "Points",
            "type": "array"
          },
          "source": {
            "default": "osm",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "points"
        ],
        "title": "Footprint",
        "type": "object"
      },
      "FootprintResponse": {
        "properties": {
          "footprint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Footprint"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "footprint"
        ],
        "title": "FootprintResponse",
        "type": "object"
      },
      "GeocodeRequest": {
        "properties": {
          "query": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "GeocodeRequest",
        "type": "object"
      },
      "GeocodeResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/GeocodeResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "GeocodeResponse",
        "type": "object"
      },
      "GeocodeResult": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          },
          "postcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postcode"
          }
        },
        "required": [
          "label",
          "lat",
          "lng"
        ],
        "title": "GeocodeResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "service",
          "version"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "LatLng": {
        "properties": {
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lng"
        ],
        "title": "LatLng",
        "type": "object"
      },
      "SunPosition": {
        "description": "Solar position in the pvlib frame: azimuth 0=N, 90=E, 180=S, 270=W.",
        "properties": {
          "azimuth": {
            "description": "Compass azimuth, degrees (0=N, 90=E, 180=S, 270=W).",
            "title": "Azimuth",
            "type": "number"
          },
          "elevation": {
            "description": "Apparent elevation above the horizon, degrees.",
            "title": "Elevation",
            "type": "number"
          }
        },
        "required": [
          "azimuth",
          "elevation"
        ],
        "title": "SunPosition",
        "type": "object"
      },
      "SurfaceYield": {
        "properties": {
          "annual_kwh": {
            "title": "Annual Kwh",
            "type": "number"
          },
          "azimuth": {
            "title": "Azimuth",
            "type": "number"
          },
          "kwp": {
            "title": "Kwp",
            "type": "number"
          },
          "specific_kwh_per_kwp": {
            "description": "PVGIS annual yield per kWp at this orientation.",
            "title": "Specific Kwh Per Kwp",
            "type": "number"
          },
          "surface_id": {
            "title": "Surface Id",
            "type": "string"
          },
          "tilt": {
            "title": "Tilt",
            "type": "number"
          }
        },
        "required": [
          "surface_id",
          "tilt",
          "azimuth",
          "kwp",
          "specific_kwh_per_kwp",
          "annual_kwh"
        ],
        "title": "SurfaceYield",
        "type": "object"
      },
      "SystemYield": {
        "properties": {
          "annual_kwh": {
            "title": "Annual Kwh",
            "type": "number"
          },
          "monthly_kwh": {
            "description": "12 values, Jan..Dec, aggregated over surfaces.",
            "items": {
              "type": "number"
            },
            "title": "Monthly Kwh",
            "type": "array"
          },
          "per_surface": {
            "items": {
              "$ref": "#/components/schemas/SurfaceYield"
            },
            "title": "Per Surface",
            "type": "array"
          }
        },
        "required": [
          "annual_kwh",
          "per_surface",
          "monthly_kwh"
        ],
        "title": "SystemYield",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "YieldRequest": {
        "properties": {
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          },
          "surfaces": {
            "items": {
              "$ref": "#/components/schemas/YieldSurfaceInput"
            },
            "title": "Surfaces",
            "type": "array"
          }
        },
        "required": [
          "lat",
          "lng",
          "surfaces"
        ],
        "title": "YieldRequest",
        "type": "object"
      },
      "YieldSurfaceInput": {
        "description": "One roof surface to evaluate: orientation + installed capacity.",
        "properties": {
          "azimuth": {
            "description": "pvlib frame: 0=N, 90=E, 180=S, 270=W.",
            "exclusiveMaximum": 360.0,
            "minimum": 0.0,
            "title": "Azimuth",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kwp": {
            "minimum": 0.0,
            "title": "Kwp",
            "type": "number"
          },
          "tilt": {
            "maximum": 90.0,
            "minimum": 0.0,
            "title": "Tilt",
            "type": "number"
          }
        },
        "required": [
          "id",
          "tilt",
          "azimuth",
          "kwp"
        ],
        "title": "YieldSurfaceInput",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Geo, solar, economics and report services for Solar Planner.",
    "title": "Solar Planner API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/geo/footprint": {
      "get": {
        "operationId": "footprint_api_v1_geo_footprint_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Lng",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FootprintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Footprint",
        "tags": [
          "geo"
        ]
      }
    },
    "/api/v1/geo/geocode": {
      "post": {
        "operationId": "geocode_api_v1_geo_geocode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeocodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeocodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Geocode",
        "tags": [
          "geo"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "description": "Liveness probe. No external dependencies are checked here.",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/solar/sun-position": {
      "get": {
        "operationId": "sun_position_api_v1_solar_sun_position_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Lng",
              "type": "number"
            }
          },
          {
            "description": "ISO 8601 instant; naive is treated as UTC.",
            "in": "query",
            "name": "timestamp",
            "required": true,
            "schema": {
              "description": "ISO 8601 instant; naive is treated as UTC.",
              "title": "Timestamp",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SunPosition"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sun Position",
        "tags": [
          "solar"
        ]
      }
    },
    "/api/v1/solar/yield": {
      "post": {
        "operationId": "system_yield_api_v1_solar_yield_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YieldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemYield"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "System Yield",
        "tags": [
          "solar"
        ]
      }
    }
  }
}